Flytter en util-funksjon dit den hører hjemme
